Move shared code to a file without go:build tags
