+++ central/splunk/violations_query_test.go
@@ -1,5 +1,3 @@
-//go:build sql_integration
-
 package splunk
 
 // This file contains datastore helpers and tests for queryAlerts() function.
@@ -9,8 +7,10 @@ import (
 	"fmt"
 	"sort"
 	"testing"
+	"time"
 
 	"github.com/blevesearch/bleve"
+	"github.com/gogo/protobuf/types"
 	"github.com/stackrox/rox/central/alert/datastore"
 	"github.com/stackrox/rox/central/globalindex"
 	"github.com/stackrox/rox/generated/storage"
@@ -19,10 +19,33 @@ import (
 	"github.com/stackrox/rox/pkg/rocksdb"
 	"github.com/stackrox/rox/pkg/sac"
 	"github.com/stackrox/rox/pkg/testutils/rocksdbtest"
+	"github.com/stackrox/rox/pkg/utils"
 	"github.com/stackrox/rox/pkg/uuid"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+	"github.com/stretchr/testify/suite"
 )
+
+// Test suite to test everything to do with splunk violations. Actual test is split across multiple files.
+// Moved here since some helpers in this file references it and for non-postgres tests, it won't be able to read it
+// if the definition is another file
+type violationsTestSuite struct {
+	suite.Suite
+	deployAlert, processAlert, k8sAlert, networkAlert, resourceAlert *storage.Alert
+	allowCtx                                                         context.Context
+}
+
+func makeTimestamp(timeStr string) *types.Timestamp {
+	ts, err := types.TimestampProto(mustParseTime(timeStr))
+	utils.CrashOnError(err)
+	return ts
+}
+
+func mustParseTime(timeStr string) time.Time {
+	ts, err := time.Parse(time.RFC3339Nano, timeStr)
+	utils.CrashOnError(err)
+	return ts
+}
 
 // testDataStore contains all things that need to be created and disposed in order to use Alerts datastore.DataStore in
 // tests.
